// apps/portal-api/src/model/kanel/public/DeploymentRequestQuota.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the table public.DeploymentRequestQuota */
export default interface DeploymentRequestQuota {
  id: string | null;

  region: string | null;

  platform_identifier: string | null;

  capacity: number | null;

  availability: number | null;
}

/** Represents the initializer for the table public.DeploymentRequestQuota */
export interface DeploymentRequestQuotaInitializer {
  /** Default value: gen_random_uuid() */
  id?: string | null;

  region?: string | null;

  platform_identifier?: string | null;

  capacity?: number | null;

  availability?: number | null;
}

/** Represents the mutator for the table public.DeploymentRequestQuota */
export interface DeploymentRequestQuotaMutator {
  id?: string | null;

  region?: string | null;

  platform_identifier?: string | null;

  capacity?: number | null;

  availability?: number | null;
}